warehouse: Consolidate and refactor errors
