+++ warehouse/templates/manage/account/webauthn-provision.html
@@ -38,11 +38,6 @@
     <label for="webauthn-provision-label" class="form-group__label">Name your device to begin</label>
     <input id="webauthn-provision-label" type="text" name="label" class="form-group__input" aria-describedby="webauthn-errors">
     <p class="form-group__help-text">PyPI supports adding multiple security devices.<br>Please give this device a name. 64 characters or fewer. All Unicode is valid, including spaces.</p>
-    <ul id="webauthn-errors" class="form-errors margin-top--large">
-      <li id="webauthn-browser-support" class="hidden">
-        <a href="https://developer.mozilla.org/en-US/docs/Web/API/PublicKeyCredential#Browser_compatibility" title="External link" target="_blank" rel="noopener">Upgrade your browser</a> to set up two factor authentication with a security device (e.g. USB key)
-      </li>
-    </ul>
   </div>
   <div>
     <button
@@ -52,6 +47,11 @@
       csrf-token="{{ request.session.get_csrf_token() }}"
       disabled>Set up security device
     </button>
+    <ul id="webauthn-errors" class="form-errors margin-top--large">
+      <li id="webauthn-browser-support" class="hidden">
+        <a href="https://developer.mozilla.org/en-US/docs/Web/API/PublicKeyCredential#Browser_compatibility" title="External link" target="_blank" rel="noopener">Upgrade your browser</a> to set up two factor authentication with a security device (e.g. USB key)
+      </li>
+    </ul>
   </div>
 
   <br>
